PersonalBudgeter/requirements.txt
Django
pillow
requests
djangorestframework
environs
